+++ src/operation/AbstractOperation.ts
@@ -1,4 +1,4 @@
-import { AbstractSerializable } from './AbstractSerializable';
+import { AbstractSerializable } from '../AbstractSerializable';
 
 export abstract class AbstractOperation extends AbstractSerializable {
   public abstract apply(sourceText: string): string;
