+++ src/operation/ConcatOperation.ts
@@ -1,6 +1,6 @@
-import { AbstractOperation } from '../AbstractOperation';
 import { fillConstantsInString } from '../fillConstantsInString';
 import { SerializableClassInstance } from '../serializable';
+import { AbstractOperation } from './AbstractOperation';
 
 export enum ConcatPosition {
   START = 'start',
